Change font managment. Now the application fellow the system font. In the application settings, we can only change the size of the button.

+++ src/ui/panel_model/scenes/AbstractItem.h
@@ -33,13 +33,13 @@ public:
 
     AbstractScene* mScene;
     
-    qreal mBorderWidth;
-    qreal mTitleHeight;
-    qreal mPhasesHeight;
-    qreal mEltsMargin;
-    qreal mEltsWidth;
-    qreal mEltsHeight;
-    qreal mItemWidth;
+    int mBorderWidth;
+    int mTitleHeight;
+    int mPhasesHeight;
+    int mEltsMargin;
+    int mEltsWidth;
+    int mEltsHeight;
+    int mItemWidth;
     
     bool mMoving; // used in AbstractScene::itemReleased() to merge item like eventItem and phaseItem
     bool mMergeable;
